feature(shared/ui): смержил UiBackdrop с UiBottomSheet
